+++ components/product-details/ProductInfo.jsx
@@ -2,59 +2,59 @@
  * This code was generated by Builder.io.
  */
 "use client"
-import React , { useEffect, useState }from "react";
+import React from "react";
 import ColorSelector from "./ColorSelector";
 import SizeSelector from "./SizeSelector";
 import QuantitySelector from "./QuantitySelector";
 import DeliveryInfo from "./DeliveryInfo";
-import ProductId from "../../api/buyer"
-
-function ProductInfo() {
 
 
-  const [productData, setProductData] = useState(null);
+function ProductInfo({productData}) {
 
-  // Lấy id từ URL
-  const urlParams = new URLSearchParams(window.location.search);
-  const productId = urlParams.get('id');
-  console.log(productId);
-  const fallbackImage =
-  "https://via.placeholder.com/150";
-  useEffect(() => {
-    const fetchProductById = async () => {
-      try {
-        const response = await ProductId.product.getByIdProduct(productId);
-        console.log(response.statusCode);
-        if (response.statusCode === 200) {
-          const productData = {
-            id: response.data.productId,
-            name: response.data.productName,
-            discount: 0,  // Lấy discount từ API hoặc gán giá trị mặc định
-            currentPrice: "", 
-            originalPrice: response.data.price,   // Nếu có giá gốc
-            rating: 5,  // Lấy rating từ API hoặc gán mặc định là 5
-            reviews: 65,  // Lấy số reviews từ API hoặc gán mặc định
-            image: fallbackImage,  // Gán ảnh mặc định nếu không có ảnh từ API
-          };
+
+  // const [productData, setProductData] = useState(null);
+
+  // // Lấy id từ URL
+  // const urlParams = new URLSearchParams(window.location.search);
+  // const productId = urlParams.get('id');
+  // console.log(productId);
+  // const fallbackImage =
+  // "https://via.placeholder.com/150";
+  // useEffect(() => {
+  //   const fetchProductById = async () => {
+  //     try {
+  //       const response = await ProductId.product.getByIdProduct(productId);
+  //       console.log(response.statusCode);
+  //       if (response.statusCode === 200) {
+  //         const productData = {
+  //           id: response.data.productId,
+  //           name: response.data.productName,
+  //           discount: 0,  // Lấy discount từ API hoặc gán giá trị mặc định
+  //           currentPrice: "", 
+  //           originalPrice: response.data.price,   // Nếu có giá gốc
+  //           rating: 5,  // Lấy rating từ API hoặc gán mặc định là 5
+  //           reviews: 65,  // Lấy số reviews từ API hoặc gán mặc định
+  //           image: fallbackImage,  // Gán ảnh mặc định nếu không có ảnh từ API
+  //         };
   
-          setProductData(productData);  // Gán dữ liệu sản phẩm vào state
-          console.log("Product data: ", productData);
-        } else {
-          console.error("Failed to fetch product details");
-        }
-      } catch (error) {
-        console.error("Error fetching product details:", error);
-      }
-    };
+  //         setProductData(productData);  // Gán dữ liệu sản phẩm vào state
+  //         console.log("Product data: ", productData);
+  //       } else {
+  //         console.error("Failed to fetch product details");
+  //       }
+  //     } catch (error) {
+  //       console.error("Error fetching product details:", error);
+  //     }
+  //   };
   
-    if (productId) {
-      fetchProductById();  // Chỉ fetch khi có productId
-    }
-  }, [productId]);
+  //   if (productId) {
+  //     fetchProductById();  // Chỉ fetch khi có productId
+  //   }
+  // }, [productId]);
   
-  if (!productData) {
-    return <div>Loading...</div>;
-  }
+  // if (!productData) {
+  //   return <div>Loading...</div>;
+  // }
 
   return (
     <div className="flex flex-col ml-5 w-[36%] max-md:ml-0 max-md:w-full">
