+++ components/product-details/ProductPage.jsx
@@ -1,19 +1,65 @@
 /**
  * This code was generated by Builder.io.
  */
-import React from "react";
+"use client"
+import React ,{ useEffect, useState } from "react";
 import Header from "./Header";
 import ProductDetails from "./ProductDetails";
 import RelatedProducts from "./RelatedProducts";
 import Footer from "./Footer";
+import ProductId from "../../api/buyer"
 
 function ProductPage() {
+  const [productData, setProductData] = useState(null);
+
+  // Lấy id từ URL
+  const urlParams = new URLSearchParams(window.location.search);
+  const productId = urlParams.get('id');
+  console.log(productId);
+  const fallbackImage =
+  "https://via.placeholder.com/150";
+  useEffect(() => {
+    const fetchProductById = async () => {
+      try {
+        const response = await ProductId.product.getByIdProduct(productId);
+        console.log(response.statusCode);
+        if (response.statusCode === 200) {
+          const productData = {
+            id: response.data.productId,
+            name: response.data.productName,
+            discount: 0,  // Lấy discount từ API hoặc gán giá trị mặc định
+            currentPrice: "", 
+            originalPrice: response.data.price,   // Nếu có giá gốc
+            rating: 5,  // Lấy rating từ API hoặc gán mặc định là 5
+            reviews: 65,  // Lấy số reviews từ API hoặc gán mặc định
+            image: fallbackImage,  // Gán ảnh mặc định nếu không có ảnh từ API
+          };
+  
+          setProductData(productData);  // Gán dữ liệu sản phẩm vào state
+          console.log("Product data: ", productData);
+        } else {
+          console.error("Failed to fetch product details");
+        }
+      } catch (error) {
+        console.error("Error fetching product details:", error);
+      }
+    };
+  
+    if (productId) {
+      fetchProductById();  // Chỉ fetch khi có productId
+    }
+  }, [productId]);
+  
+  if (!productData) {
+    return <div>Loading...</div>;
+  }
+
   return (
     <div className="flex overflow-hidden flex-col bg-white">
       <Header />
       <div className="flex justify-center items-center flex-1">
         <main>
-          <ProductDetails />
+          <ProductDetails productData={productData}/>
           <RelatedProducts />
         </main>
       </div>
